refactor: tests utils reorganise
